% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcompoplot.R
\name{char2pal}
\alias{char2pal}
\title{Creates a named color palette.}
\usage{
char2pal(x, pal = rainbow)
}
\arguments{
\item{x}{a vector of identifiers to be used for colors.}

\item{pal}{a color palette. Default is \code{\link[grDevices]{rainbow}}}
}
\value{
a named character vector of hexadecimal colors.
}
\description{
This is useful for defining a color palette that can be used by population factors.
}
\examples{
char2pal(LETTERS)
}
