% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcompoplot.R
\name{ggcompoplot}
\alias{ggcompoplot}
\title{Plot posterior values from DAPC analysis in adegenet}
\usage{
ggcompoplot(da.object, gid, pal = rainbow, cols = 1)
}
\arguments{
\item{da.object}{an object of class "dapc"}

\item{gid}{an object of class "genind"}

\item{pal}{a color palette}

\item{cols}{the number of columns to display}
}
\value{
a ggplot object with each population stacked on top of each other.
}
\description{
Plot posterior values from DAPC analysis in adegenet
}
\examples{
library('adegenet')
library('ggcompoplot')
library('ggplot2')
data(microbov)
strata(microbov) <- data.frame(other(microbov))
dapc1 <- dapc(microbov, n.pca=20, n.da=15)
setPop(microbov) <- ~breed
compoplot(dapc1, lab="") # Adegenet compoplot
# Showing per breed
ggcompoplot(dapc1, microbov) + theme(axis.text.x = element_blank())
\dontrun{

# 3 columns
ggcompoplot(dapc1, microbov, col = 3) + theme(axis.text.x = element_blank())

# Different color palette
ggcompoplot(dapc1, microbov, col = 3, pal = funky) + theme(axis.text.x = element_blank())

# Per Country
setPop(microbov) <- ~coun
ggcompoplot(dapc1, microbov) + theme(axis.text.x = element_blank())
}
}
